refactor: change argument from userId to id and add findByEmail method

+++ src/module/users/users.service.ts
@@ -1,13 +1,18 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, ConflictException, Injectable } from '@nestjs/common';
 import { UserRepository } from './repositories/users.repository';
 import { CreateUserDTO } from './dto/create-user.dto';
 import { UpdateUserDTO } from './dto/update-user.dto';
+import { User } from './entities/user.entity';
 
 @Injectable()
 export class UsersService {
   constructor(private usersRepository: UserRepository) {}
 
-  async create(userData: CreateUserDTO) {
+  async create(userData: CreateUserDTO): Promise<User> {
+
+    const existingUser = await this.usersRepository.findByEmail(userData.email)
+    if(existingUser) throw new ConflictException('Email already exists')
+
     return this.usersRepository.create(userData)
   }
 
@@ -16,16 +21,26 @@ export class UsersService {
   }
 
   async findOne(id: string) {
-    return this.usersRepository.findOne(id)
+    const existingUser = await this.usersRepository.findOne(id)
+    if(!existingUser) throw new BadRequestException('User not found')
+
+    return existingUser
   }
 
   async update(id: string, updatedData: UpdateUserDTO) {
+
+    const existingUser = await this.usersRepository.findOne(id)
+    if(!existingUser) throw new BadRequestException('User not found')
+    
     return this.usersRepository.update(id, updatedData)
   }
 
   async delete(id: string) {
-    return this.usersRepository.delete(id)
 
+    const existingUser = await this.usersRepository.findOne(id)
+    if(!existingUser) throw new BadRequestException('User not found')
+
+    return this.usersRepository.delete(id)
   }
 
 }
